restructure / added metadata
